feat(modding): added offset option for teleporter
Also fixed a bunch of issues with teleporter+worm tail interaction
