+++ Starship/Assets/Scripts/Gui/Windows/AnimatedWindow.cs
